created buffer file change print f to count for spaces
the buffer files was sperated to make the functinno printf file
to look clean and the printf file now contains a space counter
