+++ holberton.h
@@ -29,5 +29,6 @@ char *int_to_string(va_list list);
 char *percent_to_string(va_list list);
 int _strlen(char *s);
 void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size);
+char *create_buffer(void);
 
 #endif /* _PRINTF */
